added GraphViewer

+++ CAL_ViagemEuropa/requests.h
@@ -25,4 +25,4 @@ using namespace Poco::JSON;
 
 void getLatLong(string const &name, double &lat, double &lon);
 
-double getTravelDuration(double const &latSrc, double const &lonSrc, double const &latDst, double const &lonDst);
+int getTravelDuration(double const &latSrc, double const &lonSrc, double const &latDst, double const &lonDst);
